+++ solarforecastarbiter/cli.py
@@ -15,10 +15,10 @@ import sentry_sdk
 
 from solarforecastarbiter import __version__
 from solarforecastarbiter.io import nwp
-from solarforecastarbiter.io.api import (request_cli_access_token, APISession,
-                                         mock_raw_report_endpoints)
+from solarforecastarbiter.io.api import request_cli_access_token, APISession
 from solarforecastarbiter.io.fetch import start_cluster
 from solarforecastarbiter.io.reference_observations import reference_data
+from solarforecastarbiter.io.utils import mock_raw_report_endpoints
 import solarforecastarbiter.reference_forecasts.main as reference_forecasts
 from solarforecastarbiter.validation import tasks as validation_tasks
 import solarforecastarbiter.reports.main as reports
