+++ solarforecastarbiter/io/api.py
@@ -1,10 +1,8 @@
 """
 Functions to connect to and process data from SolarForecastArbiter API
 """
-from contextlib import contextmanager
 import json
 import logging
-import re
 
 
 import requests
@@ -1187,73 +1185,3 @@ class APISession(requests.Session):
         """
         req = self.get('/users/current')
         return req.json()
-
-
-@contextmanager
-def mock_raw_report_endpoints(base_url):
-    """
-    Mock API report endpoints under base_url to enable testing
-    of the report generation task run via the dashboard. Requires
-    requests_mock>=1.8.0. Catches all report endpoints, but if
-    report generation requires POSTing to other endpoints in the
-    future, they will need to be added here.
-    """
-    import requests_mock
-    value_dict = {}  # for raw processed values
-
-    def post_value_callback(request, context):
-        context.status_code = 200
-        rjson = request.json()
-        id_ = rjson['object_id']
-        value_dict[id_] = rjson['processed_values']
-        return id_
-
-    def get_value_callback(request, context):
-        context.status_code = 200
-        return [{'id': k, 'processed_values': v}
-                for k, v in value_dict.items()]
-
-    raw_dict = {}  # for the raw reports
-
-    def post_raw_callback(request, context):
-        context.status_code = 200
-        raw_dict.update(request.json())
-
-    def get_raw_callback(request, context):
-        context.status_code = 200
-        return raw_dict
-
-    report_dict = {}  # for new and full reports
-
-    def post_report_callback(request, context):
-        context.status_code = 200
-        report_dict.update(request.json())
-        return 'no_id'
-
-    def get_report_callback(request, context):
-        context.status_code = 200
-        out = report_dict.copy()
-        if raw_dict:
-            out['raw_report'] = raw_dict
-            out['values'] = [
-                {'id': k, 'processed_values': v}
-                for k, v in value_dict.items()]
-        out['status'] = 'complete'
-        return out
-
-    with requests_mock.Mocker(real_http=True) as m:
-        value_re = re.compile(f'{base_url}/reports/.*/values')
-        raw_re = re.compile(f'{base_url}/reports/.*/raw')
-        m.register_uri('GET', re.compile(
-            f'{base_url}/reports/.*'),
-                       json=get_report_callback)
-        m.register_uri('POST', re.compile(
-            f'{base_url}/reports/*'),
-                       text=post_report_callback)
-        m.register_uri('POST', re.compile(
-            f'{base_url}/reports/.*/status/.*'))
-        m.register_uri('POST', value_re, text=post_value_callback)
-        m.register_uri('GET', value_re, json=get_value_callback)
-        m.register_uri('POST', raw_re, json=post_raw_callback)
-        m.register_uri('GET', raw_re, json=get_raw_callback)
-        yield
